requirements.txt
streamlit>=1.30.0
yfinance>=0.2.30
xgboost>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
openpyxl>=3.1.0
joblib>=1.3.0
playwright>=1.40.0
